+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## dbt-core 1.0.0 (Release TBD)
 
+### Under the hood
+- Fix intermittent errors in partial parsing tests ([#4060](https://github.com/dbt-labs/dbt-core/issues/4060), [#4068](https://github.com/dbt-labs/dbt-core/pull/4068))
+
 ## dbt-core 1.0.0b1 (October 11, 2021)
 
 ## dbt 1.0.0b1 (October 11, 2021)